Rename file formable-page.vue to FormablePage.vue
