+++ pages/formable-pages/pix-certification-application.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script>
-import FormablePage from '~/components/formable-page'
+import FormablePage from '@/components/FormablePage'
 
 export default {
   name: 'PixCertificationApplication',
